+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 7.1.0 (in progress)
 
 * [#3225](https://github.com/oshi/oshi/pull/3225): Add CgroupInfo API with Linux cgroup v1/v2 support - [@rohan-coder02](https://github.com/rohan-coder02).
+* [#3223](https://github.com/oshi/oshi/issues/3223): Add `oshi-metrics` module with Micrometer integration for system metrics following OpenTelemetry semantic conventions - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here!
 
 # 7.0.0 (2026-04-30), 7.0.1 (2026-05-02)